+++ tests/testthat/test-capping.R
@@ -151,9 +151,8 @@ cap.irace <- function(...)
   irace:::checkTargetFiles(scenario = scenario, parameters = parameters)
   
   confs <- irace(scenario = scenario, parameters = parameters)
-  best.conf <- getFinalElites(logFile = scenario$logFile, n = 1,
-                              drop.metadata = TRUE)
-  expect_identical(removeConfigurationsMetaData(confs[1, , drop = FALSE]),
+  best.conf <- getFinalElites(scenario$logFile, n = 1L, drop.metadata = TRUE)
+  expect_identical(removeConfigurationsMetaData(confs[1L, , drop = FALSE]),
                    best.conf)
 }