+++ tests/testthat/test-sann-irace.R
@@ -75,8 +75,7 @@ sann.irace <- function(log.param=FALSE, ...)
   scenario <- checkScenario (scenario)
 
   confs <- irace(scenario = scenario, parameters = parameters)
-  best.conf <- getFinalElites(logFile = scenario$logFile, n = 1,
-                              drop.metadata = TRUE)
+  best.conf <- getFinalElites(scenario$logFile, n = 1, drop.metadata = TRUE)
   expect_identical(removeConfigurationsMetaData(confs[1, , drop = FALSE]),
                    best.conf)
 }